% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/applyLocalSearch.R
\name{applyLocalSearch}
\alias{applyLocalSearch}
\title{Given an individual and an instance, this function performs
local search (focused on tour length minimization) accounting
for partial tours which are already driven.}
\usage{
applyLocalSearch(ind, instance, more.args, ...)
}
\arguments{
\item{ind}{[\code{VRPIndividual}]\cr
Individual.}

\item{instance}{[\code{Network}]\cr
Netgen network.}

\item{more.args}{[\code{list}]\cr
Further parameter passed down to local search procedure.}

\item{...}{[any]\cr
Not used at the moment.}
}
\value{
Modified individual.
}
\description{
Given an individual and an instance, this function performs
local search (focused on tour length minimization) accounting
for partial tours which are already driven.
}
