% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeTourLength.R
\name{computeTourLength}
\alias{computeTourLength}
\title{Helper for tour length computation.}
\usage{
computeTourLength(ind, instance, vehicle = 1L)
}
\arguments{
\item{ind}{[\code{VRPIndividual}]\cr
Instance of type \code{VRPIndividual}.}

\item{instance}{[\code{Network}\cr
Netgen network with slots \dQuote{coords} and \dQuote{dmat}, i.e., coordinates
and distance matrix with start and end depot!}

\item{vehicle}{[\code{integer(1)}]\cr
ID of vehicle.}
}
\value{
[\code{numeric(1)}]
}
\description{
Helper for tour length computation.
}
\seealso{
fitness.fun
}
