% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_funs.R
\name{decideRandom}
\alias{decideRandom}
\title{Decision makers.}
\usage{
decideRandom(fitness, ...)
}
\arguments{
\item{fitness}{[matrix]\cr
Fitness matrix where each column represents one non-dominated solution.}

\item{...}{[any]\cr
Not used at the moment.}

\item{q}{[\numeric(1)]\cr
Numeric values in \eqn{[0, 1]}. Used by \code{decideRank}.}
}
\value{
[integer(1)] Index, i.e., column number, of solution to decide for.
}
\description{
Decision maker selects a solution uniformly at random.
}
