% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{findFixedTour}
\alias{findFixedTour}
\title{Find tour prefix of already visited customers.}
\usage{
findFixedTour(ind, instance, time.bound)
}
\arguments{
\item{ind}{[\code{VRPIndividual}]\cr
Individual.}

\item{instance}{[\code{Network}]\cr
Network object.}

\item{time.bound}{[\code{numeric(1)}]\cr
Current time. This is relevant since the prefix tour is composed
of all customers which already are served at a given point in time.}
}
\value{
[\code{integer}]
}
\description{
Find tour prefix of already visited customers.
}
