% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getListOfToursByEras.R
\name{getListOfToursByEras}
\alias{getListOfToursByEras}
\title{Extract list of decision maker tours and init tours from EMOA result.}
\usage{
getListOfToursByEras(result, eras = NULL)
}
\arguments{
\item{result}{[\code{list}]\cr
Return value of \code{\link{dynamicVRPEMOA}}.}

\item{eras}{[\code{integer}]\cr
Which eras to consider.
Default is all eras.}
}
\value{
[\code{list}] List with components \dQuote{dm.tours} and \dQuote{init.tours}
as needed by \code{\link{plotNetworkFancy}}.
}
\description{
Extract list of decision maker tours and init tours from EMOA result.
}
