% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{getTourFromIndividual}
\alias{getTourFromIndividual}
\title{Construct tour from individual.}
\usage{
getTourFromIndividual(ind, ...)
}
\arguments{
\item{ind}{[\code{VRPIndividual}]\cr
Individual.}

\item{...}{[any]\cr
Currently not used.}
}
\description{
Construct tour from individual.
}
