% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{initIndividual}
\alias{initIndividual}
\title{Initialize individual.}
\usage{
initIndividual(instance, current.time = 0, init.tours = integer(),
  n.vehicles = 1L, template.ind = NULL, init.distribution = "binomial")
}
\arguments{
\item{instance}{[\code{\link[salesperson]{Network}}]\cr
Network instance.}

\item{current.time}{[\code{numeric(1)}]\cr
Current point in time.
Default is 0.}

\item{init.tour}{[\code{integer}]\cr
Fixed prefix tour, i.e., part of tour which is already fixed,
because time passed and vehicle already visited some customers.}

\item{template.tour}{[\code{integer}]\cr
Tour used as a \dQuote{template} for a newly generated individual. Here,
we aim to pass as much information from \code{template.tour} as possible.}
}
\value{
[\code{VRPIndividual}] List with following components:
\describe{
  \item[\code{b}]{Binary vector of length |V| - 2. b[i] is 1, if customer i is active, i.e., in tour.}
  \item[\code{t}]{Permutation vector.}
  \item[\code{p}]{Vector of mutation probablities. I.e., p[i] is the probability to flip b[i]. p[i] is zero if customer i is already fixed or not yet released.}
  \item[\code{it}]{Binary vector of length |V| - 2. it[i] is 1 if and only if customer i is in initial tour.}
  \item[n.mandatory \code{integer(1)}]{Number of mandatory customers.}
  \item[idx.dynamic.available \code{integer}]{IDs/positions of dynamic customers which already requested serving.}
  \item[n.dynamic.active \code{integer(1)}]{Number of active dynamic customers i (i.e., b[i] = 1)}
  \item[n.dynamic.inactive \code{integer(1)}]{Number of available, but not active dynamic customers i (i.e., b[i] = 0)}
  \item[init.tours \code{integer}]{Fixed tour part, i.e., sequence of nodes already visited.}
}
}
