% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutation.R
\name{mutVRP}
\alias{mutVRP}
\title{Mutation operator.}
\usage{
mutVRP(ind, p.swap = 1)
}
\description{
Performs two consecutive mutation steps:
1) Mutate the binary string ind$b. I.e., dynamic, available, but not yet fixed customers
may be added or removed from customer list.
2) Swap mutation on ind$b, i.e., reorder the customer permutation.
}
