% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualization.R
\name{plotNetworkFancy}
\alias{plotNetworkFancy}
\title{Visualize tours driven by vehicles by era.}
\usage{
plotNetworkFancy(instance, time.resolution, tours = NULL,
  init.tours = NULL, customers.by.era = TRUE,
  highlight.depots = TRUE, desaturate.nonvisited = FALSE,
  filter.eras = NULL, facet.type = "grid", facet.args = list())
}
\arguments{
\item{instance}{[\code{Network}]\cr
Network.}

\item{time.resolution}{[\code{numeric(1)}]\cr
Time resolution used by \code{\link{dynamicVRPEMOA}}.}

\item{tours}{[\code{list}]\cr
List of lists of integer vectors. Each top.level list represents an era
and the second-level lists contain the tours of the vehicles.}

\item{init.tours}{[\code{list}]\cr
See documentation of \code{tours} for the structure. Contains for each
era to depict the tours already traveled by the vehicles until the
beginning of the corresponding era.}

\item{customers.by.era}{[\code{logical(1)}]\cr
Colour customers by the era they arrived at?
Default is \code{TRUE}.}

\item{highlight.depots}{[\code{logical(1)}]\cr
Show start and end depot as large white circles with black border?
Default is \code{TRUE}.}

\item{desaturate.nonvisited}{[\code{logical(1)}]\cr
This one is experimental.
Shall non-visited customers by depicted with reduced opacity?
Default is \code{FALSE}.}
}
\value{
[\code{\link[ggplot2]{ggplot}}]
}
\description{
Visualize tours driven by vehicles by era.
}
