% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{print.VRPIndividual}
\alias{print.VRPIndividual}
\title{Simple printer for individuals.}
\usage{
\method{print}{VRPIndividual}(x, ...)
}
\arguments{
\item{x}{[\code{VRPIndividual}]\cr
Individual.}

\item{...}{[any]\cr
Currently not used.}
}
\description{
Simple printer for individuals.
}
