% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision_funs.R
\name{decide}
\alias{decide}
\alias{decideRandom}
\alias{decideInteractive}
\alias{decideRank}
\title{Decision maker}
\usage{
decideRandom(fitness, ...)

decideInteractive(fitness, ...)

decideRank(obj, q = 0.5, ...)
}
\arguments{
\item{fitness}{[matrix]\cr
Fitness matrix where each column represents one non-dominated solution.}

\item{...}{[any]\cr
Not used at the moment.}

\item{q}{[\numeric(1)]\cr
Numeric values in \eqn{[0, 1]}. Used by \code{decideRank}.}
}
\value{
[integer(1)] Index, i.e., column number, of solution to decide for.
}
\description{
Function to simulate a decision maker. Function \code{decideRandom}
selects a solution uniformly at random. Function \code{decideInteractive} pauses
the execution and asks the user for manual selection in the terminal. Function
\code{decideRank} performs an order-based selection, i.e., one selects an objective
and a quantile and the corresponding elements ordered by the selected objective
is chosen.
}
