% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findHamiltonianPath.R
\name{findHamiltonianPath}
\alias{findHamiltonianPath}
\title{@export}
\usage{
findHamiltonianPath(instance,
  active.nodes = seq_len(getNumberOfNodes(instance) + 2L), start.id = 80L,
  dest.id = 2L)
}
\description{
@export
}
