% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{getToursFromIndividual}
\alias{getToursFromIndividual}
\title{Construct tour from individual.}
\usage{
getToursFromIndividual(ind, append.depots = FALSE, ...)
}
\arguments{
\item{ind}{[\code{VRPIndividual}]\cr
Individual.}

\item{...}{[any]\cr
Currently not used.}
}
\description{
Construct tour from individual.
}
