% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/individual.R
\name{initIndividual}
\alias{initIndividual}
\title{Initialize individual.}
\usage{
initIndividual(instance, current.time = 0, init.tour = integer())
}
\arguments{
\item{instance}{[\code{\link[salesperson]{Network}}]\cr
Network instance.}

\item{current.time}{[\code{numeric(1)}]\cr
Current point in time.
Default is 0.}

\item{init.tour}{[\code{integer}]\cr
Fixed prefix tour, i.e., part of tour which is already fixed,
because time passed and vehicle already visited some customers.}
}
\value{
[\code{VRPIndividual}] List with following components:
\describe{
  \item[\code{b}]{Binary vector of length |V| - 2. b[i] is 1, if customer i is active, i.e., in tour.}
  \item[\code{t}]{Permutation vector.}
  \item[\code{p}]{Vector of mutation probablities. I.e., p[i] is the probability to flip b[i]. p[i] is zero if customer i is already fixed or not yet released.}
  \item[n.mandatory \code{integer(1)}]{Number of mandatory customers.}
  \item[idx.dynamic.available \code{integer}]{IDs/positions of dynamic customers which already requested serving.}
  \item[init.tour \code{integer}]{Fixed tour part, i.e., sequence of nodes already visited.}
}
}
